requirements.txt
fastapi
uvicorn[standard]
pypdf
httpx
numpy
python-multipart
python-dotenv
